Move e2e test files into subpackage
Before I add the docker test code from https://github.com/GoogleCloudPlatform/opentelemetry-operations-java/pull/364 to this repo, I wanted to move the e2e test runner code (which is separate to a subpackage). This makes it more explicit what is used for what and simplifies CI a bit.
